Fix timeline AI state access boundary
